feat: adding language selector on userinfo
